requirement.txt
numpy==1.20.3
Pillow==8.2.0
scipy==1.7.0
